poly-event-sniper/pyproject.toml
[project]
name = "poly-event-sniper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.2",
    "aiosqlite>=0.20.0",
    "feedparser>=6.0.12",
    "loguru>=0.7.3",
    "py-clob-client>=0.34.0",
    "pydantic-settings>=2.12.0",
    "textual>=6.11.0",
    "web3>=7.14.0",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "types-aiofiles>=25.1.0.20251011",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = "test_*.py"
